Style: 커스텀 예외 처리 패키지 구성 및 매개변수 이름 수정 #66
