--- src/main/java/com/nhnacademy/marketgg/server/exception/AssetNotFoundException.java
@@ -1,9 +1,0 @@
-package com.nhnacademy.marketgg.server.exception;
-
-public class AssetNotFoundException extends IllegalArgumentException {
-
-    public AssetNotFoundException(String ex) {
-        super(ex);
-    }
-
-}
