--- src/main/java/com/nhnacademy/marketgg/server/exception/CartNotFoundException.java
@@ -1,9 +1,0 @@
-package com.nhnacademy.marketgg.server.exception;
-
-public class CartNotFoundException extends IllegalArgumentException {
-
-    public CartNotFoundException(String ex) {
-        super(ex);
-    }
-
-}
