--- src/main/java/com/nhnacademy/marketgg/server/exception/CategorizationNotFoundException.java
@@ -1,9 +1,0 @@
-package com.nhnacademy.marketgg.server.exception;
-
-public class CategorizationNotFoundException extends IllegalArgumentException {
-
-    public CategorizationNotFoundException(String s) {
-        super(s);
-    }
-
-}
