--- src/main/java/com/nhnacademy/marketgg/server/exception/CategoryNotFoundException.java
@@ -1,9 +1,0 @@
-package com.nhnacademy.marketgg.server.exception;
-
-public class CategoryNotFoundException extends IllegalArgumentException {
-
-    public CategoryNotFoundException(String message) {
-        super(message);
-    }
-
-}
