--- src/main/java/com/nhnacademy/marketgg/server/exception/CouponNotFoundException.java
@@ -1,9 +1,0 @@
-package com.nhnacademy.marketgg.server.exception;
-
-public class CouponNotFoundException extends IllegalArgumentException {
-
-    public CouponNotFoundException(String ex) {
-        super(ex);
-    }
-
-}
