--- src/main/java/com/nhnacademy/marketgg/server/exception/CrowdfundingNotFoundException.java
@@ -1,8 +1,0 @@
-package com.nhnacademy.marketgg.server.exception;
-
-public class CrowdfundingNotFoundException extends IllegalArgumentException {
-
-    public CrowdfundingNotFoundException(String ex) {
-        super(ex);
-    }
-}
