--- src/main/java/com/nhnacademy/marketgg/server/exception/CustomerServicePostNotFoundException.java
@@ -1,9 +1,0 @@
-package com.nhnacademy.marketgg.server.exception;
-
-public class CustomerServicePostNotFoundException extends IllegalArgumentException {
-
-    public CustomerServicePostNotFoundException(String ex) {
-        super(ex);
-    }
-
-}
