--- src/main/java/com/nhnacademy/marketgg/server/exception/DeliveryAddressNotFoundException.java
@@ -1,8 +1,0 @@
-package com.nhnacademy.marketgg.server.exception;
-
-public class DeliveryAddressNotFoundException extends IllegalArgumentException {
-
-    public DeliveryAddressNotFoundException(String ex) {
-        super(ex);
-    }
-}
