--- src/main/java/com/nhnacademy/marketgg/server/exception/DibNotFoundException.java
@@ -1,9 +1,0 @@
-package com.nhnacademy.marketgg.server.exception;
-
-public class DibNotFoundException extends IllegalArgumentException {
-
-    public DibNotFoundException(String ex) {
-        super(ex);
-    }
-
-}
