--- src/main/java/com/nhnacademy/marketgg/server/exception/FundingProductNotFoundException.java
@@ -1,9 +1,0 @@
-package com.nhnacademy.marketgg.server.exception;
-
-public class FundingProductNotFoundException extends IllegalArgumentException {
-
-    public FundingProductNotFoundException(String ex) {
-        super(ex);
-    }
-
-}
