--- src/main/java/com/nhnacademy/marketgg/server/exception/GivenCouponNotFoundException.java
@@ -1,8 +1,0 @@
-package com.nhnacademy.marketgg.server.exception;
-
-public class GivenCouponNotFoundException extends IllegalArgumentException {
-
-    public GivenCouponNotFoundException(String ex) {
-        super(ex);
-    }
-}
