--- src/main/java/com/nhnacademy/marketgg/server/exception/ImageNotFoundException.java
@@ -1,9 +1,0 @@
-package com.nhnacademy.marketgg.server.exception;
-
-public class ImageNotFoundException extends IllegalArgumentException {
-
-    public ImageNotFoundException(String ex) {
-        super(ex);
-    }
-
-}
