--- src/main/java/com/nhnacademy/marketgg/server/exception/LabelNotFoundException.java
@@ -1,9 +1,0 @@
-package com.nhnacademy.marketgg.server.exception;
-
-public class LabelNotFoundException extends IllegalArgumentException {
-
-    public LabelNotFoundException(String ex) {
-        super(ex);
-    }
-
-}
