--- src/main/java/com/nhnacademy/marketgg/server/exception/MemberGradeNotFoundException.java
@@ -1,8 +1,0 @@
-package com.nhnacademy.marketgg.server.exception;
-
-public class MemberGradeNotFoundException extends IllegalArgumentException {
-
-    public MemberGradeNotFoundException(String ex) {
-        super(ex);
-    }
-}
