--- src/main/java/com/nhnacademy/marketgg/server/exception/MemberNotFoundException.java
@@ -1,9 +1,0 @@
-package com.nhnacademy.marketgg.server.exception;
-
-public class MemberNotFoundException extends IllegalArgumentException {
-
-    public MemberNotFoundException(String ex) {
-        super(ex);
-    }
-
-}
