--- src/main/java/com/nhnacademy/marketgg/server/exception/OrderDeliveryAddressNotFoundException.java
@@ -1,8 +1,0 @@
-package com.nhnacademy.marketgg.server.exception;
-
-public class OrderDeliveryAddressNotFoundException extends IllegalArgumentException {
-
-    public OrderDeliveryAddressNotFoundException(String ex) {
-        super(ex);
-    }
-}
