--- src/main/java/com/nhnacademy/marketgg/server/exception/OrderNotFoundException.java
@@ -1,8 +1,0 @@
-package com.nhnacademy.marketgg.server.exception;
-
-public class OrderNotFoundException extends IllegalArgumentException {
-
-    public OrderNotFoundException(String ex) {
-        super(ex);
-    }
-}
