--- src/main/java/com/nhnacademy/marketgg/server/exception/OrderProductNotFoundException.java
@@ -1,9 +1,0 @@
-package com.nhnacademy.marketgg.server.exception;
-
-public class OrderProductNotFoundException extends IllegalArgumentException {
-
-    public OrderProductNotFoundException(String ex) {
-        super(ex);
-    }
-
-}
