--- src/main/java/com/nhnacademy/marketgg/server/exception/PaymentNotFoundException.java
@@ -1,9 +1,0 @@
-package com.nhnacademy.marketgg.server.exception;
-
-public class PaymentNotFoundException extends IllegalArgumentException {
-
-    public PaymentNotFoundException(String ex) {
-        super(ex);
-    }
-
-}
