--- src/main/java/com/nhnacademy/marketgg/server/exception/PointHistoryNotFoundException.java
@@ -1,9 +1,0 @@
-package com.nhnacademy.marketgg.server.exception;
-
-public class PointHistoryNotFoundException extends IllegalArgumentException {
-
-    public PointHistoryNotFoundException(String ex) {
-        super(ex);
-    }
-
-}
