--- src/main/java/com/nhnacademy/marketgg/server/exception/ProductInquiryPostNotFoundException.java
@@ -1,9 +1,0 @@
-package com.nhnacademy.marketgg.server.exception;
-
-public class ProductInquiryPostNotFoundException extends IllegalArgumentException {
-
-    public ProductInquiryPostNotFoundException(String ex) {
-        super(ex);
-    }
-
-}
