--- src/main/java/com/nhnacademy/marketgg/server/exception/ProductLabelNotFoundException.java
@@ -1,9 +1,0 @@
-package com.nhnacademy.marketgg.server.exception;
-
-public class ProductLabelNotFoundException extends IllegalArgumentException {
-
-    public ProductLabelNotFoundException(String ex) {
-        super(ex);
-    }
-
-}
