--- src/main/java/com/nhnacademy/marketgg/server/exception/ProductNotFoundException.java
@@ -1,8 +1,0 @@
-package com.nhnacademy.marketgg.server.exception;
-
-public class ProductNotFoundException extends RuntimeException{
-    public ProductNotFoundException(String message) {
-        super(message);
-    }
-
-}
