--- src/main/java/com/nhnacademy/marketgg/server/exception/ReviewNotFoundException.java
@@ -1,9 +1,0 @@
-package com.nhnacademy.marketgg.server.exception;
-
-public class ReviewNotFoundException extends IllegalArgumentException {
-
-    public ReviewNotFoundException(String ex) {
-        super(ex);
-    }
-
-}
