--- src/main/java/com/nhnacademy/marketgg/server/exception/TransactionNotFoundException.java
@@ -1,9 +1,0 @@
-package com.nhnacademy.marketgg.server.exception;
-
-public class TransactionNotFoundException extends IllegalArgumentException {
-
-    public TransactionNotFoundException(String ex) {
-        super(ex);
-    }
-
-}
