--- src/main/java/com/nhnacademy/marketgg/server/exception/UsedCouponNotFoundException.java
@@ -1,8 +1,0 @@
-package com.nhnacademy.marketgg.server.exception;
-
-public class UsedCouponNotFoundException extends IllegalArgumentException {
-
-    public UsedCouponNotFoundException(String ex) {
-        super(ex);
-    }
-}
