+++ src/main/java/com/nhnacademy/marketgg/server/exception/asset/AssetNotFoundException.java
@@ -1,0 +1,9 @@
+package com.nhnacademy.marketgg.server.exception.asset;
+
+public class AssetNotFoundException extends IllegalArgumentException {
+
+    public AssetNotFoundException(String msg) {
+        super(msg);
+    }
+
+}
