+++ src/main/java/com/nhnacademy/marketgg/server/exception/cart/CartNotFoundException.java
@@ -1,0 +1,9 @@
+package com.nhnacademy.marketgg.server.exception.cart;
+
+public class CartNotFoundException extends IllegalArgumentException {
+
+    public CartNotFoundException(String msg) {
+        super(msg);
+    }
+
+}
