+++ src/main/java/com/nhnacademy/marketgg/server/exception/categorization/CategorizationNotFoundException.java
@@ -1,0 +1,9 @@
+package com.nhnacademy.marketgg.server.exception.categorization;
+
+public class CategorizationNotFoundException extends IllegalArgumentException {
+
+    public CategorizationNotFoundException(String msg) {
+        super(msg);
+    }
+
+}
