+++ src/main/java/com/nhnacademy/marketgg/server/exception/category/CategoryNotFoundException.java
@@ -1,0 +1,9 @@
+package com.nhnacademy.marketgg.server.exception.category;
+
+public class CategoryNotFoundException extends IllegalArgumentException {
+
+    public CategoryNotFoundException(String msg) {
+        super(msg);
+    }
+
+}
