+++ src/main/java/com/nhnacademy/marketgg/server/exception/coupon/CouponNotFoundException.java
@@ -1,0 +1,9 @@
+package com.nhnacademy.marketgg.server.exception.coupon;
+
+public class CouponNotFoundException extends IllegalArgumentException {
+
+    public CouponNotFoundException(String msg) {
+        super(msg);
+    }
+
+}
