+++ src/main/java/com/nhnacademy/marketgg/server/exception/crowdfunding/CrowdfundingNotFoundException.java
@@ -1,0 +1,9 @@
+package com.nhnacademy.marketgg.server.exception.crowdfunding;
+
+public class CrowdfundingNotFoundException extends IllegalArgumentException {
+
+    public CrowdfundingNotFoundException(String msg) {
+        super(msg);
+    }
+
+}
