+++ src/main/java/com/nhnacademy/marketgg/server/exception/customerservicepost/CustomerServicePostNotFoundException.java
@@ -1,0 +1,9 @@
+package com.nhnacademy.marketgg.server.exception.customerservicepost;
+
+public class CustomerServicePostNotFoundException extends IllegalArgumentException {
+
+    public CustomerServicePostNotFoundException(String msg) {
+        super(msg);
+    }
+
+}
