+++ src/main/java/com/nhnacademy/marketgg/server/exception/deliveryaddresses/DeliveryAddressNotFoundException.java
@@ -1,0 +1,9 @@
+package com.nhnacademy.marketgg.server.exception.deliveryaddresses;
+
+public class DeliveryAddressNotFoundException extends IllegalArgumentException {
+
+    public DeliveryAddressNotFoundException(String msg) {
+        super(msg);
+    }
+
+}
