+++ src/main/java/com/nhnacademy/marketgg/server/exception/dib/DibNotFoundException.java
@@ -1,0 +1,9 @@
+package com.nhnacademy.marketgg.server.exception.dib;
+
+public class DibNotFoundException extends IllegalArgumentException {
+
+    public DibNotFoundException(String msg) {
+        super(msg);
+    }
+
+}
