+++ src/main/java/com/nhnacademy/marketgg/server/exception/fundingproduct/FundingProductNotFoundException.java
@@ -1,0 +1,9 @@
+package com.nhnacademy.marketgg.server.exception.fundingproduct;
+
+public class FundingProductNotFoundException extends IllegalArgumentException {
+
+    public FundingProductNotFoundException(String msg) {
+        super(msg);
+    }
+
+}
