+++ src/main/java/com/nhnacademy/marketgg/server/exception/givencoupon/GivenCouponNotFoundException.java
@@ -1,0 +1,9 @@
+package com.nhnacademy.marketgg.server.exception.givencoupon;
+
+public class GivenCouponNotFoundException extends IllegalArgumentException {
+
+    public GivenCouponNotFoundException(String msg) {
+        super(msg);
+    }
+
+}
