+++ src/main/java/com/nhnacademy/marketgg/server/exception/image/ImageNotFoundException.java
@@ -1,0 +1,9 @@
+package com.nhnacademy.marketgg.server.exception.image;
+
+public class ImageNotFoundException extends IllegalArgumentException {
+
+    public ImageNotFoundException(String msg) {
+        super(msg);
+    }
+
+}
