+++ src/main/java/com/nhnacademy/marketgg/server/exception/label/LabelNotFoundException.java
@@ -1,0 +1,9 @@
+package com.nhnacademy.marketgg.server.exception.label;
+
+public class LabelNotFoundException extends IllegalArgumentException {
+
+    public LabelNotFoundException(String msg) {
+        super(msg);
+    }
+
+}
