+++ src/main/java/com/nhnacademy/marketgg/server/exception/member/MemberNotFoundException.java
@@ -1,0 +1,9 @@
+package com.nhnacademy.marketgg.server.exception.member;
+
+public class MemberNotFoundException extends IllegalArgumentException {
+
+    public MemberNotFoundException(String msg) {
+        super(msg);
+    }
+
+}
