+++ src/main/java/com/nhnacademy/marketgg/server/exception/membergrade/MemberGradeNotFoundException.java
@@ -1,0 +1,9 @@
+package com.nhnacademy.marketgg.server.exception.membergrade;
+
+public class MemberGradeNotFoundException extends IllegalArgumentException {
+
+    public MemberGradeNotFoundException(String msg) {
+        super(msg);
+    }
+
+}
