+++ src/main/java/com/nhnacademy/marketgg/server/exception/order/OrderNotFoundException.java
@@ -1,0 +1,9 @@
+package com.nhnacademy.marketgg.server.exception.order;
+
+public class OrderNotFoundException extends IllegalArgumentException {
+
+    public OrderNotFoundException(String msg) {
+        super(msg);
+    }
+
+}
