+++ src/main/java/com/nhnacademy/marketgg/server/exception/orderdeliveryaddress/OrderDeliveryAddressNotFoundException.java
@@ -1,0 +1,9 @@
+package com.nhnacademy.marketgg.server.exception.orderdeliveryaddress;
+
+public class OrderDeliveryAddressNotFoundException extends IllegalArgumentException {
+
+    public OrderDeliveryAddressNotFoundException(String msg) {
+        super(msg);
+    }
+
+}
