+++ src/main/java/com/nhnacademy/marketgg/server/exception/orderproduct/OrderProductNotFoundException.java
@@ -1,0 +1,9 @@
+package com.nhnacademy.marketgg.server.exception.orderproduct;
+
+public class OrderProductNotFoundException extends IllegalArgumentException {
+
+    public OrderProductNotFoundException(String msg) {
+        super(msg);
+    }
+
+}
