+++ src/main/java/com/nhnacademy/marketgg/server/exception/payment/PaymentNotFoundException.java
@@ -1,0 +1,9 @@
+package com.nhnacademy.marketgg.server.exception.payment;
+
+public class PaymentNotFoundException extends IllegalArgumentException {
+
+    public PaymentNotFoundException(String msg) {
+        super(msg);
+    }
+
+}
