+++ src/main/java/com/nhnacademy/marketgg/server/exception/pointhistory/PointHistoryNotFoundException.java
@@ -1,0 +1,9 @@
+package com.nhnacademy.marketgg.server.exception.pointhistory;
+
+public class PointHistoryNotFoundException extends IllegalArgumentException {
+
+    public PointHistoryNotFoundException(String msg) {
+        super(msg);
+    }
+
+}
