+++ src/main/java/com/nhnacademy/marketgg/server/exception/product/ProductNotFoundException.java
@@ -1,0 +1,9 @@
+package com.nhnacademy.marketgg.server.exception.product;
+
+public class ProductNotFoundException extends RuntimeException{
+
+    public ProductNotFoundException(String msg) {
+        super(msg);
+    }
+
+}
