+++ src/main/java/com/nhnacademy/marketgg/server/exception/productinquiry/ProductInquiryPostNotFoundException.java
@@ -1,0 +1,9 @@
+package com.nhnacademy.marketgg.server.exception.productinquiry;
+
+public class ProductInquiryPostNotFoundException extends IllegalArgumentException {
+
+    public ProductInquiryPostNotFoundException(String ex) {
+        super(ex);
+    }
+
+}
