+++ src/main/java/com/nhnacademy/marketgg/server/exception/productlabel/ProductLabelNotFoundException.java
@@ -1,0 +1,9 @@
+package com.nhnacademy.marketgg.server.exception.productlabel;
+
+public class ProductLabelNotFoundException extends IllegalArgumentException {
+
+    public ProductLabelNotFoundException(String msg) {
+        super(msg);
+    }
+
+}
