+++ src/main/java/com/nhnacademy/marketgg/server/exception/review/ReviewNotFoundException.java
@@ -1,0 +1,9 @@
+package com.nhnacademy.marketgg.server.exception.review;
+
+public class ReviewNotFoundException extends IllegalArgumentException {
+
+    public ReviewNotFoundException(String msg) {
+        super(msg);
+    }
+
+}
