+++ src/main/java/com/nhnacademy/marketgg/server/exception/transaction/TransactionNotFoundException.java
@@ -1,0 +1,9 @@
+package com.nhnacademy.marketgg.server.exception.transaction;
+
+public class TransactionNotFoundException extends IllegalArgumentException {
+
+    public TransactionNotFoundException(String msg) {
+        super(msg);
+    }
+
+}
