+++ src/main/java/com/nhnacademy/marketgg/server/exception/usedcoupon/UsedCouponNotFoundException.java
@@ -1,0 +1,9 @@
+package com.nhnacademy.marketgg.server.exception.usedcoupon;
+
+public class UsedCouponNotFoundException extends IllegalArgumentException {
+
+    public UsedCouponNotFoundException(String msg) {
+        super(msg);
+    }
+
+}
