+++ src/main/java/com/nhnacademy/marketgg/server/service/impl/DefaultLabelService.java
@@ -3,7 +3,7 @@ package com.nhnacademy.marketgg.server.service.impl;
 import com.nhnacademy.marketgg.server.dto.request.LabelCreateRequest;
 import com.nhnacademy.marketgg.server.dto.response.LabelRetrieveResponse;
 import com.nhnacademy.marketgg.server.entity.Label;
-import com.nhnacademy.marketgg.server.exception.LabelNotFoundException;
+import com.nhnacademy.marketgg.server.exception.label.LabelNotFoundException;
 import com.nhnacademy.marketgg.server.repository.LabelRepository;
 import com.nhnacademy.marketgg.server.service.LabelService;
 import lombok.RequiredArgsConstructor;
