+++ src/main/java/com/nhnacademy/marketgg/server/service/impl/DefaultProductService.java
@@ -7,9 +7,9 @@ import com.nhnacademy.marketgg.server.entity.Asset;
 import com.nhnacademy.marketgg.server.entity.Category;
 import com.nhnacademy.marketgg.server.entity.Image;
 import com.nhnacademy.marketgg.server.entity.Product;
-import com.nhnacademy.marketgg.server.exception.AssetNotFoundException;
-import com.nhnacademy.marketgg.server.exception.CategoryNotFoundException;
-import com.nhnacademy.marketgg.server.exception.ProductNotFoundException;
+import com.nhnacademy.marketgg.server.exception.asset.AssetNotFoundException;
+import com.nhnacademy.marketgg.server.exception.category.CategoryNotFoundException;
+import com.nhnacademy.marketgg.server.exception.product.ProductNotFoundException;
 import com.nhnacademy.marketgg.server.repository.AssetRepository;
 import com.nhnacademy.marketgg.server.repository.CategoryRepository;
 import com.nhnacademy.marketgg.server.repository.ImageRepository;
