+++ src/test/java/com/nhnacademy/marketgg/server/service/impl/DefaultCategoryServiceTest.java
@@ -6,8 +6,8 @@ import com.nhnacademy.marketgg.server.dto.response.CategoryRetrieveResponse;
 import com.nhnacademy.marketgg.server.dto.request.CategoryUpdateRequest;
 import com.nhnacademy.marketgg.server.entity.Categorization;
 import com.nhnacademy.marketgg.server.entity.Category;
-import com.nhnacademy.marketgg.server.exception.CategorizationNotFoundException;
-import com.nhnacademy.marketgg.server.exception.CategoryNotFoundException;
+import com.nhnacademy.marketgg.server.exception.categorization.CategorizationNotFoundException;
+import com.nhnacademy.marketgg.server.exception.category.CategoryNotFoundException;
 import com.nhnacademy.marketgg.server.repository.CategorizationRepository;
 import com.nhnacademy.marketgg.server.repository.CategoryRepository;
 import com.nhnacademy.marketgg.server.service.CategoryService;
